+++ src/commands/page.ts
@@ -105,25 +105,25 @@ export async function rename(pageName: string) {
   const currentPage = await logseq.Editor.getCurrentPage();
   if (currentPage) {
     await logseq.Editor.renamePage(currentPage.name, pageName);
-    logseq.App.showMsg(`Page renamed to ${pageName}`);
+    logseq.UI.showMsg(`Page renamed to ${pageName}`);
   } else {
-    logseq.App.showMsg("Rename command only work on a page.");
+    logseq.UI.showMsg("Rename command only work on a page.");
   }
 }
 
 export function write() {
-  logseq.App.showMsg("Actually Logseq save your info automatically!");
+  logseq.UI.showMsg("Actually Logseq save your info automatically!");
 }
 
 export function writeAndQuit() {
-  logseq.App.showMsg(
+  logseq.UI.showMsg(
     "Actually Logseq save your info automatically! So just quit VIM command mode."
   );
   hideMainUI();
 }
 
 export function quit() {
-  logseq.App.showMsg("Quit VIM command mode.");
+  logseq.UI.showMsg("Quit VIM command mode.");
   hideMainUI();
 }
 
@@ -137,17 +137,17 @@ export async function substituteBlock(value) {
     const block = await logseq.Editor.getCurrentBlock();
     if (block && block.uuid && block.content) {
       await replaceBlock(block, regex, replace);
-      await logseq.App.showMsg(
+      await logseq.UI.showMsg(
         'Current block replaced "' + search + '" with "' + replace + '"'
       );
       hideMainUI();
     } else {
-      await logseq.App.showMsg(
+      await logseq.UI.showMsg(
         "Current block not found. Please select a block first."
       );
     }
   } else {
-    await logseq.App.showMsg('Please input "s/search/replace/modifiers"');
+    await logseq.UI.showMsg('Please input "s/search/replace/modifiers"');
   }
 }
 
@@ -161,15 +161,15 @@ export async function substitutePage(value) {
       const modifiers = splitReplace[3] || "";
       const regex = new RegExp(search, modifiers);
       await walkReplace(blocks, regex, replace);
-      await logseq.App.showMsg(
+      await logseq.UI.showMsg(
         'Current page blocks replaced "' + search + '" with "' + replace + '"'
       );
       hideMainUI();
     } else {
-      await logseq.App.showMsg('Please input "%s/search/replace/modifiers"');
+      await logseq.UI.showMsg('Please input "%s/search/replace/modifiers"');
     }
   } else {
-    await logseq.App.showMsg(
+    await logseq.UI.showMsg(
       "Current page blocks not found. Please select a page first."
     );
   }